Fixing python error in lsh_matmul_op_test; renaming lsh_matmul to tensorflow_lsh_matmul
