+++ tensorflow_lsh_matmul/python/lsh_keras.py
